+++ src/util/mod.rs
@@ -1,1 +1,4 @@
 pub mod err;
+pub mod opt_locale_loc;
+
+pub use opt_locale_loc::OptLocaleLoc;