+++ CHANGELOG.md
@@ -2,6 +2,13 @@ TBD
 ===
 Unreleased changes. Release notes have not yet been written.
 
+0.1.1 (2021-03-04)
+=====
+
+* add --locale <loc> to options
+* add -?, --query <q> to options
+* change output plain numeric to locale numeric, fancy numeric format
+
 0.1.0 (2021-03-03)
 =====
 first commit